Add roles to Spring Security and define User Attendance relationship
Users can have an Organizer or User role.

Regular users will be able to mark attendance for events. This requires
a new relationship between Events and Users, as well as the ability to
mark attendance in the EventService and UserService.

This commit does not add the functionality to the view yet.